backend/requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
bcrypt==3.2.0
bson==0.5.10
certifi==2020.11.8
cffi==1.14.4
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
jsonfield==3.1.0
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.20
pymongo==3.7.2
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
six==1.15.0
sqlparse==0.2.4
wincertstore==0.2
